+++ CHANGELOG.md
@@ -24,6 +24,8 @@ BUG FIXES:
   - core: Downloads with user/password use the curl `-u` flag. [GH-3183]
   - core: `vagrant help` no longer loads the Vagrantfile. [GH-3180]
   - guests/darwin: Fix an exception when configuring networks. [GH-3143]
+  - guests/linux: Only chown folders/files in rsync if they don't
+    have the proper owner. [GH-3186]
   - hosts/linux: Unusual sed delimiter to avoid conflicts. [GH-3167]
   - providers/virtualbox: Make more internal interactions with VBoxManage
     retryable to avoid spurious VirtualBox errors. [GH-2831]